recipeViewPage integrations
