+++ backend/controllers/resetDatabaseController.js
@@ -1,13 +1,13 @@
-import { User } from '../models/user.js'
-import { UserRecipe } from '../models/useCreatedreceipe.js'
+//import { User } from '../models/user.js'
+import { Recipe } from '../models/recipe.js'
 
 async function resetDatabase(request, response) {
   try {
     // Delete all records from the User collection
-    await User.deleteMany({})
+    //await User.deleteMany({})
 
     // Delete all records from the UserRecipe collection
-    await UserRecipe.deleteMany({})
+    await Recipe.deleteMany({})
 
     console.log('Database reset successful.')
     response.json({ succes: 'Database reset successful.' })
