+++ backend/models/recipe.js
@@ -5,7 +5,6 @@ import mongoose from 'mongoose'
 const recipeSchema = new mongoose.Schema({
   recipeId: {
     type: String,
-    required: true,
   },
   label: {
     type: String,
@@ -14,6 +13,25 @@ const recipeSchema = new mongoose.Schema({
   image: {
     type: String,
     required: true,
+  },
+  images: {
+    type: Array,
+    default: [],
+  },
+  ingredients: {
+    type: Array,
+    default: [],
+  },
+  healthLabels: {
+    type: Array,
+    default: [],
+  },
+  creator: {
+    type: mongoose.Schema.Types.ObjectId,
+    ref: 'User',
+  },
+  instructions: {
+    type: String,
   },
   likes: [
     {
