+++ backend/routes/interactionRouter.js
@@ -4,6 +4,7 @@ const interactionRouter = express.Router()
 
 interactionRouter.get('/all', interactionController.getAllInteractionRecipes)
 interactionRouter.get('/', interactionController.getAllInteractions)
+interactionRouter.get('/userCreated',interactionController.getUserCreatedInteractions)
 interactionRouter.post('/', interactionController.createInteraction)
 interactionRouter.post('/likes', interactionController.addLikeInteraction)
 interactionRouter.delete('/likes', interactionController.removeLikeInteraction)
