+++ foodapp/src/components/UserRecipeSearch.js
@@ -1,15 +1,20 @@
 import { Container, Typography, OutlinedInput, CircularProgress, Grid } from '@mui/material'
-import { useState } from 'react'
-import { useGetUserRecipesQuery } from '../services/userRecipeApiSlice'
+import { useState, useEffect } from 'react'
+import { useGetAllUserCreatedInteractionsQuery } from '../services/interactionSlice'
 import RecipeCard from './RecipeCard'
 
 
 const UserSearchPage = () => {
   const [search, setSearch] = useState('')
-  const { data: recipedata, isLoading /*,isSuccess*/, isError } = useGetUserRecipesQuery()
+  const { data: recipedata, isLoading, isError } = useGetAllUserCreatedInteractionsQuery()
 
-  console.log(recipedata)
+  useEffect(() => {
+    console.log(recipedata)
+  },[recipedata])
 
+  if(!recipedata){
+    return <div>Loading...</div>
+  }
 
 
   return (
@@ -27,7 +32,7 @@ const UserSearchPage = () => {
         alignItems="center"
       >
         {isLoading ? <CircularProgress/> : isError ? <h3>error occured while trying to get recipes</h3>
-          : recipedata.map((recipe) =>  <RecipeCard key={recipe.id} recipe={recipe} />) }
+          : recipedata.recipesCreatedByUsrs.map((recipe) =>  <RecipeCard key={recipe.id} recipe={recipe} />) }
       </Grid>
     </Container>
   )
