+++ foodapp/src/services/interactionSlice.js
@@ -47,6 +47,14 @@ export const interactionApiSlice = apiSlice.injectEndpoints({
       },
       providesTags: ['Interactions']
     }),
+    getAllUserCreatedInteractions: builder.query({
+      query: () => {
+        return {
+          url: '/interactions/userCreated',
+        }
+      },
+      providesTags: ['Interactions']
+    }),
     createInteraction: builder.mutation({
       query: ({ recipeId, label, image }) => {
         return ({
@@ -75,5 +83,6 @@ export const {
   useRemoveDislikeInteractionMutation,
   useGetAllInteractionsQuery,
   useCreateInteractionMutation,
-  useGetAllInteractionRecipesQuery
+  useGetAllInteractionRecipesQuery,
+  useGetAllUserCreatedInteractionsQuery,
 } = interactionApiSlice
